Renamed project so it is automatically identified as unit test project
