+++ dotnet-monitor.sln
@@ -48,7 +48,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Diagnostics.Monit
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Diagnostics.Monitoring.Tool.UnitTestCommon", "src\Tests\Microsoft.Diagnostics.Monitoring.Tool.UnitTestCommon\Microsoft.Diagnostics.Monitoring.Tool.UnitTestCommon.csproj", "{E438C0E0-0301-4A8F-9153-BFE4839F213E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CollectionRuleActionUnitTests", "src\Tests\CollectionRuleActionUnitTests\CollectionRuleActionUnitTests.csproj", "{46B22C1B-1934-44E1-8BF5-3D18A1EE9163}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CollectionRuleActions.UnitTests", "src\Tests\CollectionRuleActions.UnitTests\CollectionRuleActions.UnitTests.csproj", "{2FD0A2A3-D7DA-4458-9084-E195102B9D5B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -124,10 +124,10 @@ Global
 		{E438C0E0-0301-4A8F-9153-BFE4839F213E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E438C0E0-0301-4A8F-9153-BFE4839F213E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E438C0E0-0301-4A8F-9153-BFE4839F213E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{46B22C1B-1934-44E1-8BF5-3D18A1EE9163}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{46B22C1B-1934-44E1-8BF5-3D18A1EE9163}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{46B22C1B-1934-44E1-8BF5-3D18A1EE9163}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{46B22C1B-1934-44E1-8BF5-3D18A1EE9163}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2FD0A2A3-D7DA-4458-9084-E195102B9D5B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2FD0A2A3-D7DA-4458-9084-E195102B9D5B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2FD0A2A3-D7DA-4458-9084-E195102B9D5B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2FD0A2A3-D7DA-4458-9084-E195102B9D5B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -152,7 +152,7 @@ Global
 		{F324BAD6-C9C0-42DE-9150-D8307D972E9A} = {C7568468-1C79-4944-8136-18812A7F9EA7}
 		{9BEB8058-7EE4-4EE2-B83D-592E85236E78} = {C7568468-1C79-4944-8136-18812A7F9EA7}
 		{E438C0E0-0301-4A8F-9153-BFE4839F213E} = {C7568468-1C79-4944-8136-18812A7F9EA7}
-		{46B22C1B-1934-44E1-8BF5-3D18A1EE9163} = {C7568468-1C79-4944-8136-18812A7F9EA7}
+		{2FD0A2A3-D7DA-4458-9084-E195102B9D5B} = {C7568468-1C79-4944-8136-18812A7F9EA7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {46465737-C938-44FC-BE1A-4CE139EBB5E0}
