+++ src/Tests/CollectionRuleActions.UnitTests/ActionListExecutorTests.cs
@@ -19,7 +19,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace CollectionRuleActionUnitTests
+namespace CollectionRuleActions.UnitTests
 {
     public sealed class ActionListExecutorTests
     {
