+++ src/Tests/CollectionRuleActions.UnitTests/CollectDumpActionTests.cs
@@ -6,44 +6,49 @@ using Microsoft.Diagnostics.Monitoring.TestCommon;
 using Microsoft.Diagnostics.Monitoring.TestCommon.Options;
 using Microsoft.Diagnostics.Monitoring.TestCommon.Runners;
 using Microsoft.Diagnostics.Monitoring.WebApi;
+using Microsoft.Diagnostics.Monitoring.WebApi.Models;
 using Microsoft.Diagnostics.Tools.Monitor.CollectionRules;
 using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Actions;
-using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Exceptions;
 using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Options.Actions;
 using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.IO;
 using System.Reflection;
-using System.Text;
-using System.Threading;
+using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace CollectionRuleActionUnitTests
+namespace CollectionRuleActions.UnitTests
 {
-    public sealed class CollectGCDumpActionTests
+    public sealed class CollectDumpActionTests
     {
-        private const string DefaultRuleName = "GCDumpTestRule";
+        private const string DefaultRuleName = nameof(CollectDumpActionTests);
 
-        private readonly ITestOutputHelper _outputHelper;
+        private ITestOutputHelper _outputHelper;
         private readonly EndpointUtilities _endpointUtilities;
 
-        public CollectGCDumpActionTests(ITestOutputHelper outputHelper)
+        public CollectDumpActionTests(ITestOutputHelper outputHelper)
         {
             _outputHelper = outputHelper;
             _endpointUtilities = new(_outputHelper);
         }
 
         [Theory]
-        [MemberData(nameof(ActionTestsHelper.GetTfms), MemberType = typeof(ActionTestsHelper))]
-        public Task CollectGCDumpAction_Success(TargetFrameworkMoniker tfm)
+        [MemberData(nameof(ActionTestsHelper.GetTfmsAndDumpTypes), MemberType = typeof(ActionTestsHelper))]
+        public Task CollectDumpAction_Success(TargetFrameworkMoniker tfm, DumpType dumpType)
         {
-            return Retry(() => CollectGCDumpAction_SuccessCore(tfm));
+            return Retry(dumpType, () => CollectDumpAction_SuccessCore(tfm, dumpType));
         }
 
-        private async Task CollectGCDumpAction_SuccessCore(TargetFrameworkMoniker tfm)
+        private async Task CollectDumpAction_SuccessCore(TargetFrameworkMoniker tfm, DumpType dumpType)
         {
+            // MacOS dumps inconsistently segfault the runtime on .NET 5: https://github.com/dotnet/dotnet-monitor/issues/174
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX) && tfm == TargetFrameworkMoniker.Net50)
+            {
+                return;
+            }
+
             using TemporaryDirectory tempDirectory = new(_outputHelper);
 
             await TestHostHelper.CreateCollectionRulesHost(_outputHelper, rootOptions =>
@@ -51,14 +56,14 @@ namespace CollectionRuleActionUnitTests
                 rootOptions.AddFileSystemEgress(ActionTestsConstants.ExpectedEgressProvider, tempDirectory.FullName);
 
                 rootOptions.CreateCollectionRule(DefaultRuleName)
-                    .AddCollectGCDumpAction(ActionTestsConstants.ExpectedEgressProvider)
+                    .AddCollectDumpAction(ActionTestsConstants.ExpectedEgressProvider, dumpType)
                     .SetStartupTrigger();
             }, async host =>
             {
-                CollectGCDumpOptions options = ActionTestsHelper.GetActionOptions<CollectGCDumpOptions>(host, DefaultRuleName);
+                CollectDumpOptions options = ActionTestsHelper.GetActionOptions<CollectDumpOptions>(host, DefaultRuleName);
 
                 ICollectionRuleActionFactoryProxy factory;
-                Assert.True(host.Services.GetService<ICollectionRuleActionOperations>().TryCreateFactory(KnownCollectionRuleActions.CollectGCDump, out factory));
+                Assert.True(host.Services.GetService<ICollectionRuleActionOperations>().TryCreateFactory(KnownCollectionRuleActions.CollectDump, out factory));
 
                 EndpointInfoSourceCallback callback = new(_outputHelper);
                 await using ServerSourceHolder sourceHolder = await _endpointUtilities.StartServerAsync(callback);
@@ -73,36 +78,26 @@ namespace CollectionRuleActionUnitTests
 
                     ICollectionRuleAction action = factory.Create(endpointInfo, options);
 
-                    CollectionRuleActionResult result = await ActionTestsHelper.ExecuteAndDisposeAsync(action, CommonTestTimeouts.GCDumpTimeout);
+                    CollectionRuleActionResult result = await ActionTestsHelper.ExecuteAndDisposeAsync(action, CommonTestTimeouts.DumpTimeout);
 
                     string egressPath = ActionTestsHelper.ValidateEgressPath(result);
 
-                    using FileStream gcdumpStream = new(egressPath, FileMode.Open, FileAccess.Read);
-                    Assert.NotNull(gcdumpStream);
+                    using FileStream dumpStream = new(egressPath, FileMode.Open, FileAccess.Read);
+                    Assert.NotNull(dumpStream);
 
-                    await ValidateGCDump(gcdumpStream);
+                    await DumpTestUtilities.ValidateDump(runner.Environment.ContainsKey(DumpTestUtilities.EnableElfDumpOnMacOS), dumpStream);
 
                     await runner.SendCommandAsync(TestAppScenarios.AsyncWait.Commands.Continue);
                 });
             });
         }
 
-        private static async Task ValidateGCDump(Stream gcdumpStream)
+        private async Task Retry(DumpType type, Func<Task> func, int attemptCount = 3)
         {
-            using CancellationTokenSource cancellation = new(CommonTestTimeouts.GCDumpTimeout);
-            byte[] buffer = await gcdumpStream.ReadBytesAsync(24, cancellation.Token);
+            bool isMacOSFullDump =
+                RuntimeInformation.IsOSPlatform(OSPlatform.OSX) &&
+                type == DumpType.Full;
 
-            const string knownHeaderText = "!FastSerialization.1";
-
-            Encoding enc8 = Encoding.UTF8;
-
-            string headerText = enc8.GetString(buffer, 4, knownHeaderText.Length);
-
-            Assert.Equal(knownHeaderText, headerText);
-        }
-
-        private async Task Retry(Func<Task> func, int attemptCount = 3)
-        {
             int attemptIteration = 0;
             while (true)
             {
@@ -114,11 +109,10 @@ namespace CollectionRuleActionUnitTests
 
                     break;
                 }
-                catch (CollectionRuleActionException ex) when (attemptIteration < attemptCount && ex.InnerException is InvalidOperationException)
+                catch (TaskCanceledException) when (attemptIteration < attemptCount && isMacOSFullDump)
                 {
-                    // GC dumps can fail to be produced from the runtime because the pipeline doesn't get the expected
-                    // start, data, and stop events. The pipeline will throw an InvalidOperationException, which is
-                    // wrapped in a CollectionRuleActionException by the action. Allow retries when this occurs.
+                    // Full dumps on MacOS sometimes take a very long time (longer than 100 seconds, the default
+                    // HttpClient timeout). Retry the test when this condition is detected.
                 }
             }
         }
