+++ src/Tests/CollectionRuleActions.UnitTests/CollectLiveMetricsActionTests.cs
@@ -19,7 +19,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace CollectionRuleActionUnitTests
+namespace CollectionRuleActions.UnitTests
 {
     public class CollectLiveMetricsActionTests
     {
