+++ src/Tests/CollectionRuleActions.UnitTests/CollectLogsActionTests.cs
@@ -30,7 +30,7 @@ using Xunit;
 using Xunit.Abstractions;
 using DisposableHelper = Microsoft.Diagnostics.Monitoring.TestCommon.DisposableHelper;
 
-namespace CollectionRuleActionUnitTests
+namespace CollectionRuleActions.UnitTests
 {
     public class CollectLogsActionTests
     {
