+++ src/Tests/CollectionRuleActions.UnitTests/CollectTraceActionTests.cs
@@ -20,7 +20,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace CollectionRuleActionUnitTests
+namespace CollectionRuleActions.UnitTests
 {
     public sealed class CollectTraceActionTests
     {
