+++ src/Tests/CollectionRuleActions.UnitTests/CollectionRuleActions.UnitTests.csproj
