+++ src/Tests/CollectionRuleActions.UnitTests/LoadProfilerActionTests.cs
@@ -21,7 +21,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace CollectionRuleActionUnitTests
+namespace CollectionRuleActions.UnitTests
 {
     public sealed class LoadProfilerActionTests
     {
